+++ src/cli/loader/loader-messages.js
@@ -11,4 +11,4 @@ export const BEFORE_RUNNING_SPECS = 'running specs';
 export const BEFORE_IMPORT_PUT_ON_SCOPE = 'importing components';
 export const BEFORE_PERSISTING_PUT_ON_SCOPE = 'persisting...';
 export const BEFORE_INSTALL_NPM_DEPENDENCIES = 'ensuring npm dependencies';
-
+export const BEFORE_EXPORT = 'exporting component';