+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- loader for export command
+
 ### Change
 
 - scope now fetch devDependencies (compiler/tester) on export